requirements.txt
opencv-python
numpy
matplotlib
pyyaml
svgwrite
ezdxf
